✨ [feat] 커스텀 훅 만들고 ProductInquiry에 연결해보기 #220
